ADDED A RESOURCE DIRECTORY
Added a resource directory to hold my media. Furthermore
I relinked my index.html with my index.css so that the media
and css were correctly linked to the index.html file. I needed
to do this as it was a previous portfolio I created and the
file structure was not the same.

[Ticket: X]